Initialize copx and copy lists as empty so no garbage data is collected.
Before I was initializing copx and copy lists as [0] because to create
the initial point on the graph you need some value (you can't pass an
empty list). This means the first piece of data would be 0, which could
be removed at the end easily. Instead I have adjusted it so the lists
contain only good data.