Change timer interval to 0 for the protocol and stream timers.
When the sampling rate of the DAQ is very fast ~1kHz, a timer interval
of 1ms was not fast enough. Setting the interval to 0 will trigger a
timeout as soon as all events in the queue are processed, which is
effectively triggering a timeout as fast as the timer can.