Decouple the graphs from the protocol widget so they run perpetually.
Once you click the record button the DataWorker is started and the
PlotWidget starts displaying real-time data. This runs independently
of the control of the ProtocolWidget, which is useful if you want to
visualize what is happening on the platform while not actually saving
any data. When you click "Collect baseline" on the ProtocolWidget you
connect the ProtocolWidget to the data_signal of the DataWorker, which
is how you get the data into the ProtocolWidget for processing.