fix: prevent trial start if data not streaming
There was a bug where you could click the Record button, which stops the
DAQ from sampling data, but then click the Start Trial button. No data
would get sent to the `protocol_widget`, which would raise a numpy
warning since it tried to mean of an empty array. Now if the user tries
to start a trial but the Record button is not checked a pop-up will
inform the user to enable the Record button first.