feat: save baseline trials
The baseline trials need to be saved. Added that functionality, which
meant altering create_csv_export() to use kwargs. You can pass
additional rows you want included as keyword arguments, with the key
being the row label and the value being the value of the row. This keeps
the create_csv_export() function dynamic to handle different use cases.
Also altered the generate_filename() to follow the new filename
convention.