fix: `standing_timer` running too long
The `standing_timer` would run for its entire duration, even when the
trial was stopped prematurely. This would cause the timer to timeout
at erroneous times, sometimes stopping in-progress trials. Now when the
Stop button is clicked there is a check to see if `standing_timer` is
active, and if it is the timer is stopped manually.